IslamicEval-BurhanAI-Public/abubakr/requirements.txt
python-dotenv==1.1.1
openai==1.98.0
tqdm
pandas==2.3.1
json-repair==0.48.0
scikit-learn==1.7.1
PyArabic==0.6.15
Whoosh==2.7.4
fuzzywuzzy==0.18.0
python-levenshtein==0.27.1
cohere
textdistance==4.6.3
qdrant-client==1.15.1
